use u64 for room_id/endpoint_id

+++ src/rtc/endpoint/mod.rs
@@ -1,22 +1,22 @@
 #[derive(Default)]
 pub struct Endpoint {
-    room_id: usize,
-    endpoint_id: usize,
+    room_id: u64,
+    endpoint_id: u64,
 }
 
 impl Endpoint {
-    pub fn new(room_id: usize, endpoint_id: usize) -> Self {
+    pub fn new(room_id: u64, endpoint_id: u64) -> Self {
         Self {
             room_id,
             endpoint_id,
         }
     }
 
-    pub fn room_id(&self) -> usize {
+    pub fn room_id(&self) -> u64 {
         self.room_id
     }
 
-    pub fn endpoint_id(&self) -> usize {
+    pub fn endpoint_id(&self) -> u64 {
         self.endpoint_id
     }
